refactor: Polish the code
