+++ src/components/DragSelection.tsx
@@ -1,4 +1,5 @@
 import { MOUSE_CLICK } from '@/constants/mouse'
+import { useAutoScroll } from '@/hooks/useScrollTheLad'
 import { DOMUtil, DOMVector } from '@/utils/dom'
 import React, { useCallback, useEffect, useRef, useState } from 'react'
 
@@ -24,7 +25,6 @@ interface Props {
   isReady: boolean
   containerRef: React.RefObject<HTMLDivElement | null> | null
   scrollRef?: React.RefObject<HTMLDivElement | null> | null
-  innerContainerRef?: React.RefObject<HTMLDivElement | null> | null
   onSelectionStart?: ({
     event,
     x,
@@ -40,7 +40,6 @@ const DragSelection: React.FC<Props> = ({
   isReady,
   containerRef,
   scrollRef = containerRef,
-  innerContainerRef = containerRef,
   onSelectionStart,
   onSelectionChange,
   onSelectionEnd,
@@ -210,58 +209,12 @@ const DragSelection: React.FC<Props> = ({
     scrollRef,
   ])
 
-  useEffect(() => {
-    if (!isAreaDragging) return
-
-    let handle = requestAnimationFrame(scrollTheLad)
-
-    return () => cancelAnimationFrame(handle)
-
-    function clamp(num: number, min: number, max: number) {
-      return Math.min(Math.max(num, min), max)
-    }
-
-    function scrollTheLad() {
-      if (!containerRef?.current || !scrollRef?.current || !dragVector) return
-
-      const currentPointer = dragVector.toTerminalPoint()
-      const containerRect = containerRef.current.getBoundingClientRect()
-
-      const offset = 100
-
-      const shouldScrollRight = containerRect.width - currentPointer.x < offset
-      const shouldScrollLeft = currentPointer.x < offset
-      const shouldScrollDown = containerRect.height - currentPointer.y < offset
-      const shouldScrollUp = currentPointer.y < offset
-
-      let left
-      if (shouldScrollRight) {
-        left = clamp(offset - containerRect.width + currentPointer.x, 0, 15)
-      } else if (shouldScrollLeft) {
-        left = -1 * clamp(offset - currentPointer.x, 0, 15)
-      } else {
-        left = undefined
-      }
-
-      let top
-      if (shouldScrollDown) {
-        top = clamp(offset - containerRect.height + currentPointer.y, 0, 15)
-      } else if (shouldScrollUp) {
-        top = -1 * clamp(offset - currentPointer.y, 0, 15)
-      } else {
-        top = undefined
-      }
-
-      if (top === undefined && left === undefined) {
-        handle = requestAnimationFrame(scrollTheLad)
-        return
-      }
-
-      scrollRef?.current?.scrollBy({ left, top })
-
-      handle = requestAnimationFrame(scrollTheLad)
-    }
-  }, [containerRef, dragVector, innerContainerRef, isAreaDragging, scrollRef])
+  useAutoScroll({
+    containerRef,
+    scrollRef,
+    isDragging: isAreaDragging,
+    dragVector,
+  })
 
   return (
     <div
